requirements.txt
mkdocs-material>=9.0.0
pyyaml>=6.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
